Add colorpicker component and improve generic connection
LBSD-980
