+++ client/app/scripts/liveblog-edit/styles/settings.scss
@@ -262,3 +262,7 @@ p.warn-block {
     clear: both;
     padding-top: 10px;
 }
+
+.justify-space-between {
+    justify-content: space-between;
+}
